Change sculk spread protection to its own flag
Same as the tree growth change, this moves it from the misc preventions
system to its own flag that the claim owner can change.
